# api/api/openapi.yaml
openapi: 3.0.3
info:
  title: EHR API
  description: |
    Headless Electronic Health Record API with dual interfaces:
    - FHIR R4-compliant endpoints under /fhir
    - Operational REST endpoints under /api/v1
  version: 0.1.0
  contact:
    name: EHR Engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - BearerAuth: []
paths:
...
  /fhir/Consent:
    get:
      tags: [Consent]
      summary: Search consents (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRPatient'
        - $ref: '#/components/parameters/FHIRStatus'
        - name: category
          in: query
          schema:
            type: string
          description: Consent category
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/FHIRBundle'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Consent]
      summary: Create consent (FHIR)
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Consent'
      responses:
        '201':
          description: Consent created
          headers:
            Location:
              schema:
                type: string
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Consent'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /fhir/DocumentReference:
    get:
      tags: [DocumentReference]
      summary: Search document references (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRPatient'
        - $ref: '#/components/parameters/FHIRStatus'
        - name: type
          in: query
          schema:
            type: string
          description: Document type code
        - $ref: '#/components/parameters/FHIRDate'
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/FHIRBundle'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [DocumentReference]
      summary: Create document reference (FHIR)
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/DocumentReference'
      responses:
        '201':
          description: DocumentReference created
          headers:
            Location:
              schema:
                type: string
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/DocumentReference'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /fhir/Composition:
    get:
      tags: [Composition]
      summary: Search compositions (FHIR)
      parameters:
        - $ref: '#/components/parameters/FHIRPatient'
        - $ref: '#/components/parameters/FHIRStatus'
        - name: type
          in: query
          schema:
            type: string
          description: Composition type code
        - $ref: '#/components/parameters/Count'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          $ref: '#/components/responses/FHIRBundle'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Composition]
      summary: Create composition (FHIR)
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/Composition'
      responses:
        '201':
          description: Composition created
          headers:
            Location:
              schema:
                type: string
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/Composition'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token. In development mode, the server accepts any
        well-formed JWT. In production, tokens are validated against
        the configured JWKS endpoint.
  parameters:
...
    FHIRPatient:
      name: patient
      in: query
      schema:
        type: string
      description: Patient reference (e.g. Patient/123)
    FHIRStatus:
      name: status
      in: query
      schema:
        type: string
      description: Resource status code
    FHIRDate:
      name: date
      in: query
      schema:
        type: string
      description: Date search (supports prefixes eq, ne, gt, lt, ge, le, sa, eb)
...
    Count:
      name: _count
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      description: Number of results per page (FHIR)
    Offset:
      name: _offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
      description: Result offset for pagination (FHIR)
...
  responses:
    FHIRBundle:
      description: FHIR Bundle (searchset)
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/Bundle'
    BadRequest:
      description: Invalid request
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
...
    InternalError:
      description: Internal server error
      content:
        application/fhir+json:
          schema:
            $ref: '#/components/schemas/OperationOutcome'
  schemas:
...
    Bundle:
      type: object
      properties:
        resourceType:
          type: string
          enum: [Bundle]
        type:
          type: string
          enum: [searchset]
        total:
          type: integer
          description: Total number of matching resources
        link:
          type: array
          items:
            type: object
            properties:
              relation:
                type: string
                enum: [self, next, previous]
              url:
                type: string
                format: uri
        entry:
          type: array
          items:
            type: object
            properties:
              fullUrl:
                type: string
                format: uri
              resource:
                type: object
                description: The FHIR resource
              search:
                type: object
                properties:
                  mode:
                    type: string
                    enum: [match, include, outcome]
    OperationOutcome:
      type: object
      required: [resourceType, issue]
      properties:
        resourceType:
          type: string
          enum: [OperationOutcome]
        issue:
          type: array
          items:
            type: object
            required: [severity, code]
            properties:
              severity:
                type: string
                enum: [fatal, error, warning, information]
              code:
                type: string
                description: Error code (not-found, invalid, etc.)
              diagnostics:
                type: string
                description: Human-readable diagnostic message
...
    Meta:
      type: object
      properties:
        versionId:
          type: string
        lastUpdated:
          type: string
          format: date-time
    Reference:
      type: object
      properties:
        reference:
          type: string
          description: Relative reference (e.g. Patient/123)
        display:
          type: string
    CodeableConcept:
      type: object
      properties:
        coding:
          type: array
          items:
            $ref: '#/components/schemas/Coding'
        text:
          type: string
    Coding:
      type: object
      properties:
        system:
          type: string
          format: uri
        code:
          type: string
        display:
          type: string
...
    Consent:
      type: object
      required: [resourceType, status, scope, category]
      properties:
        resourceType:
          type: string
          enum: [Consent]
        id:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'
        status:
          type: string
          enum: [draft, proposed, active, rejected, inactive, entered-in-error]
        scope:
          $ref: '#/components/schemas/CodeableConcept'
        category:
          type: array
          items:
            $ref: '#/components/schemas/CodeableConcept'
        patient:
          $ref: '#/components/schemas/Reference'
        dateTime:
          type: string
          format: date-time
    DocumentReference:
      type: object
      required: [resourceType, status, content]
      properties:
        resourceType:
          type: string
          enum: [DocumentReference]
        id:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'
        status:
          type: string
          enum: [current, superseded, entered-in-error]
        type:
          $ref: '#/components/schemas/CodeableConcept'
        subject:
          $ref: '#/components/schemas/Reference'
        date:
          type: string
          format: date-time
        content:
          type: array
          items:
            type: object
            properties:
              attachment:
                type: object
                properties:
                  contentType:
                    type: string
                  url:
                    type: string
                    format: uri
                  title:
                    type: string
    Composition:
      type: object
      required: [resourceType, status, type, subject, date, author]
      properties:
        resourceType:
          type: string
          enum: [Composition]
        id:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'
        status:
          type: string
          enum: [preliminary, final, amended, entered-in-error]
        type:
          $ref: '#/components/schemas/CodeableConcept'
        subject:
          $ref: '#/components/schemas/Reference'
        date:
          type: string
          format: date-time
        author:
          type: array
          items:
            $ref: '#/components/schemas/Reference'
        title:
          type: string
        section:
          type: array
          items:
            type: object
            properties:
              title:
                type: string
              code:
                $ref: '#/components/schemas/CodeableConcept'
              text:
                type: object
                properties:
                  div:
                    type: string
              entry:
                type: array
                items:
                  $ref: '#/components/schemas/Reference'

    # ── Operational Schemas ────────────────────────────────────────
...